Main Feature Complete
The API Endpoint for selecting jokes from a specific category doesnt seem to work. I even had to use axiom instead of fetch because I thought it was on my side. I have completed the app; it allows the user to select a category from a drop down and on submit it would have printed out the appropriate joke/fact but I improvised and printed out a random joke/fact. This joke/fact is saved to the store upon each submit.